// PolicyCOCCheck/COCPPCheck/models/IPPCOCMember.ts
export interface IPPCOCMember {

    rowNumber: number;

    relation: string;

    gender: string;

    dateOfBirth: string;

    category: string;

    maritalStatus: string;

    requiresMedicalUnderwriting: boolean;

    pfNo: string;

    memberNo: string;

    memberName: string;

    dependency: string;

    effectiveDate: string;

    previousCoverageDate: string;

    emiratesId: string;

    visaType: string;

    visaEmirate: string;

    classNo: string;

    nationality: string;

    passportNo: string;

    uidNo: string;

    mobileNo: string;

    emailId: string;

    cocEligible: boolean;

    ppEligible: boolean;

    recordStatus: string;

    validationErrors: string[];
}