// PolicyCOCCheck/COCPPCheck/models/IPPCOCRow.ts
export interface IPPCOCRow {

    rowNumber: number;

    pfNo: string;

    memberName: string;

    dateOfBirth: string;

    visaEmirate: string;

    visaType: string;

    ppCheckStatus: string;

    cocCheckStatus: string;

    prRemarks: string;

    cocAmount?: number;

    apiResponse?: any;
}